+++ demos/ids-contextual-action-panel/index.js
@@ -1,0 +1,1 @@
+import IdsContextualActionPanel from '../../src/components/ids-contextual-action-panel';